+++ tests/test_taskgroups.py
@@ -4,7 +4,7 @@ import trio
 
 from anyio import (
     create_task_group, sleep, move_on_after, fail_after, open_cancel_scope, wait_all_tasks_blocked,
-    current_effective_deadline, current_time)
+    current_effective_deadline, current_time, get_cancelled_exc_class)
 from anyio._backends import asyncio
 from anyio.exceptions import ExceptionGroup
 
@@ -417,3 +417,16 @@ async def test_timeout_error_with_multiple_cancellations():
             async with create_task_group() as tg:
                 await tg.spawn(sleep, 2)
                 await sleep(2)
+
+
+@pytest.mark.anyio
+async def test_catch_cancellation():
+    finalizer_done = False
+    async with move_on_after(0.1):
+        try:
+            await sleep(1)
+        except get_cancelled_exc_class():
+            finalizer_done = True
+            raise
+
+    assert finalizer_done